% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearspacing.R
\name{linearspacing}
\alias{linearspacing}
\title{Returns a new value for an object in an attempt to space it out from others in a vector}
\usage{
linearspacing(label, nodes, radius = 0.5, shift = 0.1, minval = -Inf,
  maxval = Inf)
}
\arguments{
\item{label}{the name of the object in the vector to consider}

\item{nodes}{the complete vector of values (with names)}

\item{radius}{how big a a radius of repulsion should we give to each value}

\item{shift}{what is the maximum amount we should shift the value}

\item{minval}{what is the minimum allowed value after shifting}

\item{maxval}{what is the maximum allowed value after shifting}
}
\description{
Returns a new value for an object in an attempt to space it out from others in a vector
}
