% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2color.R
\name{map2color}
\alias{map2color}
\title{converts numbers to colors}
\usage{
map2color(x = NULL, pal = colorRampPalette(c("deepskyblue2", "black",
  "gold"))(100), limits = NULL)
}
\arguments{
\item{x}{vector of numbers}

\item{pal}{color palette}

\item{limits}{vector of length two describing the limits of the color mapping}
}
\description{
converts numbers to colors
}
\examples{
x = 1:10
y = rep(1,10)
pal = colorRampPalette(c("red","green","blue"))(10)
cols = map2color(x,pal=pal)
plot(x=x,y=y,col=cols,pch=19)
}
