% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimizeoverlap.R
\name{minimizeoverlap}
\alias{minimizeoverlap}
\title{Define a function that iteratively shift values in a vector avoid overlap}
\usage{
minimizeoverlap(x, maxiterations = 100, radius = 0.5, shift = 0.1,
  minval = -Inf, maxval = Inf)
}
\arguments{
\item{x}{vector of values to be spaced out}

\item{maxiterations}{the maximum number of iterations to try}

\item{radius}{how big a a radius of repulsion should we give to each value}

\item{shift}{what is the maximum amount we should shift the value}

\item{minval}{what is the minimum allowed value after shifting}

\item{maxval}{what is the maximum allowed value after shifting}
}
\description{
Define a function that iteratively shift values in a vector avoid overlap
}
\examples{
minval = 0 - 0.5
maxval = 20 + 0.5
plot(1,type='n',ann=F,xaxt='n',yaxt='n',xlim=c(0,40),ylim=c(0,20),xaxs='i',yaxs='i')

# pick vals between 1 and 100
x = runif(10,0,20)
y = minimizeoverlap(x,maxiterations=100,radius=.5,shift=0.1,minval=minval,maxval=maxval)

for (i in 1:length(x))
{
  plotCircle(1,x[i],.5)
  plotCircle(2,y[i],.5)
}

}
