% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotmap.R
\name{hotmap}
\alias{hotmap}
\title{plots hotmap}
\usage{
hotmap(ourmatrix, colors, gaps = NULL, gapsize = 0.025)
}
\arguments{
\item{ourmatrix}{matrix to biuld hotmap from}

\item{colors}{colors to use for heatmap}
}
\description{
plots hotmap
}
\examples{

x = c(10,20,30,40,50,60)
mat = matrix( data =  x,nrow = 3, byrow= FALSE)
hotmap(mat)
}
